backend/requirements.txt
# ═══════════════════════════════════════
# FakeBuster AI — Python Dependencies
# ═══════════════════════════════════════

# ── Web Framework ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# ── Database ──
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# ── Redis ──
redis[hiredis]==5.2.1

# ── Auth & Security ──
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# ── HTTP Client (URL ingestion) ──
httpx==0.28.1

# ── Virus Scanning ──
clamd==1.0.2

# ── Validation & Utilities ──
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.1.1
python-dotenv==1.0.1
uuid6==2024.7.10

# ── Image Processing ──
Pillow==11.1.0

# ── Testing ──
pytest==8.3.4
pytest-asyncio==0.25.2
aiosqlite==0.20.0
